+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# **1.7.0**
+* Rework the structure of logging.
+* Tweak the launch method in the README. Should be easier to develop on. I suggest VSCode.
+* Connect logging to Application Insights. Improve logging (including source shard and human readable event type).
+* No messages, message IDs, user IDs, Guild IDs, etc are logged.
+* Add a heartbeat from each shard. Logs some metrics each hour to confirm the shard is up and connected.
+
+# **1.6.0**
+* Make bot deferral mechanism work on a per-channel basis, rather than per-server.
+
 # **1.5.5**
 * Embed help links with `<link>` syntax, to avoid embedding URLs.
 * Rollem has a dedicated support server/channel now at <https://goo.gl/7qfUtG>.